feat(makejob):完善微服务改造的修复

+++ app/question/internal/data/question_set_repo.go
@@ -102,7 +102,7 @@ func (r *questionSetRepo) GetQuestions(ctx context.Context, setID uint64) ([]*bi
 			Difficulty:   m.Difficulty,
 			Type:         m.Type,
 			IndustryCode: m.IndustryCode,
-			CategoryID:   m.CategoryID,
+			CategoryID:   uint64(m.CategoryID),
 		}
 	}
 	return questions, nil
